4/8/2024 6:25
Set up the overall structure for the different pages, the login stuff, and the navigate stuff. Backend people you still have to link the backend to this, and double check that this is working. Cause how these parts are exactly written depends on how your backend is done.

At the moment running into an error, it's saying how my react hooks aren't named correctly when they should be. Check group chat for details or npm start it yourself to see