4/9/2024 1:35pm
Homepage front-end is now done, including the looping system to output all the friends. BUT it still needs to be linked to the backend for it to work. Whoever does that please also double check that I didn't make any errors, I can't test it when it's not linked to backend.

Now homepage only has these things left:
- Casual and Ranked button needs to navigate to the respective pages and pass on the data that the matchmaking page needs
- homepage still needs to be linked to backend