4/9/2024 5:58pm
Set up basic file and code structure for casual_matchmaking page.

Modal menu for view game records in homepage is work in progress. Backend people will have to link it with the backend so that it properly pulls the game records from user info.

Also temporarily hid generateFriendsList at line 155 because since the backend hasn't been written yet it's generating an error, and I can't test the rest of the code with it doing that. If you start working on the backend for that feel free to remove the comment brackets.